refactor probestore into dedicated package
